Add -branch suffix so branch and tag do not have a name clash
A name clash causes confusing behaviour. If you have a tag and branch
called "my-name", it's not immediately clear what `git checkout my-name`
does. In fact, it checks out the branch but in our case we probably
want it to check out the tag.

Adding a `-branch` suffix helps solve this problem.