Replace Discussion stubs with new recommend signature + types
The previous stub was unused. This adds relevant types and signatures for the
`recommend` method.

This PR is based off commits in #111. This change adds `recommend` by
itself so that it can be implemented soon by Android and iOS.

We can release the remaining Discussion methods in another Bridget
version, so the platforms can implement them when ready. This will
enable us to render more articles with discussion using DCAR, more
quickly.

Co-authored-by: Jamie B <53781962+JamieB-gu@users.noreply.github.com>
Co-authored-by: Ioanna Kokkini <ioannakok@users.noreply.github.com>
Co-authored-by: Charlotte <charlotte.emms@theguardian.com>